+++ modules/fbs-core/web/src/app/page-components/course-preview/course-preview-w-progress.component.html
@@ -1,0 +1,14 @@
+<mat-card class="card" (click)="goToCourse()">
+  <span>{{ data.name }}</span
+  ><mat-icon class="invisible" *ngIf="!data.visible">visibility_off</mat-icon>
+  <span
+    [matTooltip]="
+      'Korrekt: ' + passed + '\nFalsch: ' + failed + '\nGesamt: ' + all
+    "
+    >Fortschritt:<span class="passed"> {{ passed }} </span>/<span
+      class="failed"
+    >
+      {{ failed }} </span
+    >/<span class="all"> {{ all }} </span></span
+  >
+</mat-card>
