+++ modules/fbs-core/web/src/app/page-components/course-preview/course-preview.component.html
@@ -1,19 +1,4 @@
 <mat-card class="card" (click)="goToCourse()">
   <span>{{ data.name }}</span
   ><mat-icon class="invisible" *ngIf="!data.visible">visibility_off</mat-icon>
-  <span
-    [matTooltip]="
-      'Korrekt: ' +
-      passed +
-      '\nFalsch: ' +
-      failed +
-      '\nGesamt: ' +
-      getNrOfAllTasks()
-    "
-    >Fortschritt:<span class="passed"> {{ passed }} </span>/<span
-      class="failed"
-    >
-      {{ failed }} </span
-    >/<span class="all"> {{ getNrOfAllTasks() }} </span></span
-  >
 </mat-card>
