+++ modules/fbs-core/web/src/app/page-components/my-courses/my-courses.component.html
@@ -10,10 +10,10 @@
     <mat-icon matSuffix>search</mat-icon>
   </mat-form-field>
 
-  <app-course-preview
+  <app-course-preview-w-progress
     *ngFor="let course of filteredCourses | async"
     [data]="course"
-  ></app-course-preview>
+  ></app-course-preview-w-progress>
 
   <span class="noElements" *ngIf="(courses | async)?.length === 0">
     Tragen Sie sich in <a routerLink="/courses/search">Kurse</a> ein um diese
